requirements.txt
geopy==2.4.1
folium==0.15.1
pandas==2.1.4
googlemaps==4.10.0
python-dotenv==1.0.0
requests==2.31.0 